renaming observer_config
