+++ src/launcher/mod.rs
@@ -18,7 +18,7 @@
 mod parser;
 mod ipc_server;
 mod app_handler;
-mod observer_config;
+mod observer_data;
 
 /// Launcher exposes API for managing applications
 #[derive(Clone)]
